requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
PyYAML>=6.0
plotly>=5.17.0
scipy>=1.11.0
statsmodels>=0.14.0
arch>=6.2.0
